codegen: split arch-dependent VM backend into mirrored arch/ tree (#643)
Previously x86-64 was the implicit default — its VM-tier asm
(vmgen/, invoker.rs, wrapper.rs, plus ~40 #[cfg(target_arch =
"x86_64")] methods inlined into codegen.rs and jit_module.rs) lived
directly under codegen/, while the aarch64 counterpart was isolated
in a single 2749-line arm64.rs. The two architectures were treated
asymmetrically.

Reorganize so both are peers under codegen/arch/, switched by
target_arch in codegen/arch.rs, with a mirrored file layout:

  arch/x86_64/   {codegen, jit_module, invoker, wrapper, vmgen}(+vmgen/)
  arch/aarch64/  {codegen, jit_module, invoker, wrapper, vmgen}

codegen.rs and jit_module.rs now hold only arch-neutral code (the
Codegen/JitModule type definitions, shared helpers, handle_error,
ErrorReturn, jit_check_stack). Every per-arch asm method became an
inherent impl method on the arch-neutral types, relocated into the
matching arch file; the redundant per-item target_arch cfgs are
dropped (the whole subtree is arch-gated at the mod). Methods called
across the codegen module tree were widened from private/pub(super)
to pub(in crate::codegen); the monoasm! macro, formerly supplied to
the whole tree by a codegen.rs glob import, is now imported directly
by the JIT subtree and each arch backend.

The JIT (codegen/jitgen, gated by cfg(jit) = x86-64 && !no-jit) is a
single implementation and stays under codegen/, unmoved.

No behavior change: pure relocation + visibility widening. Verified
`cargo check` on x86-64 (JIT) and aarch64 (VM-only, cross), and ran
sample programs on both (aarch64 under qemu) with identical output.
