+++ monoruby/src/codegen.rs
@@ -1,43 +1,16 @@
 use std::hash::Hash;
 
 use monoasm::*;
-#[cfg(target_arch = "x86_64")]
+// Provides the `monoasm!` macro to the JIT subtree (`compiler`, `jitgen`,
+// `patch`) via descendant visibility. The arch backends under `arch/` import
+// it directly instead.
+#[cfg(jit)]
 use monoasm_macro::monoasm;
-#[cfg(target_arch = "x86_64")]
-use paste::paste;
 #[cfg(jit)]
 use std::time::Duration;
 
-/// Integer comparison helpers (`icmp_eq` … `icmp_ge`): set `rax` to the Ruby
-/// boolean `Value` for `rdi <cond> rsi`. These are VM-tier — emitted by the
-/// bytecode handlers (`vmgen`'s `vm_*_opt_rr`) as well as the JIT — so they
-/// live here rather than in `codegen::jitgen` (which is JIT-only and gets
-/// `#[cfg]`-excluded from the aarch64 / no-jit VM build).
-#[cfg(target_arch = "x86_64")]
-macro_rules! icmp_main {
-    ($op:ident) => {
-        paste! {
-            pub(in crate::codegen) fn [<icmp_ $op>](&mut self) {
-                monoasm! { &mut self.jit,
-                    xorq rax, rax;
-                    cmpq rdi, rsi;
-                    [<set $op>] rax;
-                    shlq rax, 3;
-                    orq rax, (FALSE_VALUE);
-                };
-            }
-        }
-    };
-    ($op1:ident, $($op2:ident),+) => {
-        icmp_main!($op1);
-        icmp_main!($($op2),+);
-    };
-}
-
 #[cfg(jit)]
 mod compiler;
-#[cfg(target_arch = "x86_64")]
-mod invoker;
 mod jit_module;
 #[cfg(jit)]
 pub mod jitgen;
@@ -45,16 +18,15 @@ pub(crate) mod signal_table;
 #[cfg(jit)]
 mod patch;
 pub mod runtime;
-#[cfg(target_arch = "x86_64")]
-mod vmgen;
-#[cfg(target_arch = "x86_64")]
-mod wrapper;
 
-// aarch64 VM backend. The x86-64 VM tier above
-// (`vmgen`/`invoker`/`wrapper` + the asm methods in `jit_module` and this
-// file) is gated to x86-64; the aarch64 counterparts live here.
-#[cfg(target_arch = "aarch64")]
-mod arm64;
+// Architecture-specific VM-tier backend. The asm-level VM construction,
+// frame setup, invokers, native-function wrappers, bytecode dispatch, and
+// the small asm helpers have separate x86-64 and aarch64 implementations,
+// selected here by `target_arch`. Both live under `arch/` with a mirrored
+// file layout (`codegen.rs`, `jit_module.rs`, `invoker.rs`, `wrapper.rs`,
+// `vmgen.rs`); only the inherent `impl Codegen` / `impl JitModule` methods
+// they provide differ per arch — the types themselves stay arch-neutral.
+mod arch;
 
 #[cfg(jit)]
 use self::jitgen::asmir::AsmEvict;
@@ -295,32 +267,6 @@ impl std::ops::DerefMut for JitModule {
 }
 
 impl JitModule {
-    /// Emit a per-signal asm stub. Each stub OR-sets its own bit into
-    /// `pending_signals` and nudges `alloc_flag` so the next GC poll
-    /// in `execute_gc` notices the signal and converts it to a Ruby
-    /// exception. `signo` is the POSIX signal number (1..32).
-    ///
-    /// The handler runs in async-signal context; the only operations
-    /// performed are a memory OR and an ADD, both of which are
-    /// async-signal-safe on x86-64. No call into Rust.
-    #[cfg(target_arch = "x86_64")]
-    pub(crate) fn signal_handler_for(
-        &mut self,
-        alloc_flag: DestLabel,
-        pending_signals: DestLabel,
-        signo: i32,
-    ) -> CodePtr {
-        debug_assert!((1..=32).contains(&signo));
-        let bit: i32 = 1 << (signo - 1);
-        let codeptr = self.jit.get_current_address();
-        monoasm! { &mut self.jit,
-            addl [rip + alloc_flag], 10;
-            orl  [rip + pending_signals], (bit);
-            ret;
-        }
-        codeptr
-    }
-
     pub(crate) fn get_address_pair(&mut self) -> (CodePtr, CodePtr) {
         assert_eq!(0, self.jit.get_page());
         let ptr0 = self.jit.get_current_address();
@@ -366,460 +312,9 @@ impl JitModule {
         Self::perf_write(info, func_name);
     }
 
-    ///
-    /// Save caller-save registers (except rax) in stack.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn save_registers(&mut self) {
-        monoasm! { &mut self.jit,
-            subq rsp, 192;
-            //movq [rsp + 176], rax;
-            movq [rsp + 168], r11;
-            movq [rsp + 160], r10;
-            movq [rsp + 152], r9;
-            movq [rsp + 144], r8;
-            movq [rsp + 136], rcx;
-            movq [rsp + 128], rdx;
-            movq [rsp + 120], rsi;
-            movq [rsp + 112], rdi;
-            movq [rsp + 104], xmm15;
-            movq [rsp + 96], xmm14;
-            movq [rsp + 88], xmm13;
-            movq [rsp + 80], xmm12;
-            movq [rsp + 72], xmm11;
-            movq [rsp + 64], xmm10;
-            movq [rsp + 56], xmm9;
-            movq [rsp + 48], xmm8;
-            movq [rsp + 40], xmm7;
-            movq [rsp + 32], xmm6;
-            movq [rsp + 24], xmm5;
-            movq [rsp + 16], xmm4;
-            movq [rsp + 8], xmm3;
-            movq [rsp + 0], xmm2;
-        }
-    }
-
-    ///
-    /// Restore caller-save registers (except rax) from stack.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn restore_registers(&mut self) {
-        monoasm! { &mut self.jit,
-            movq xmm2, [rsp + 0];
-            movq xmm3, [rsp + 8];
-            movq xmm4, [rsp + 16];
-            movq xmm5, [rsp + 24];
-            movq xmm6, [rsp + 32];
-            movq xmm7, [rsp + 40];
-            movq xmm8, [rsp + 48];
-            movq xmm9, [rsp + 56];
-            movq xmm10, [rsp + 64];
-            movq xmm11, [rsp + 72];
-            movq xmm12, [rsp + 80];
-            movq xmm13, [rsp + 88];
-            movq xmm14, [rsp + 96];
-            movq xmm15, [rsp + 104];
-            movq rdi, [rsp + 112];
-            movq rsi, [rsp + 120];
-            movq rdx, [rsp + 128];
-            movq rcx, [rsp + 136];
-            movq r8, [rsp + 144];
-            movq r9, [rsp + 152];
-            movq r10, [rsp + 160];
-            movq r11, [rsp + 168];
-            //movq rax, [rsp + 176];
-            addq rsp, 192;
-        }
-    }
-
-    ///
-    /// Test whether the current local frame has been "captured" —
-    /// either already promoted to heap (bit 7 `on_heap`) or still at
-    /// its stack address but content has been moved to heap with a
-    /// tombstone (bit 3 `invalidated`). Both cases violate the JIT's
-    /// stack-layout assumptions for register-cached locals / the
-    /// specialised rbp-relative outer access and require a deopt.
-    ///
-    /// The `invalidated` case arises when a method call (often an
-    /// inlined wrapper like the Ruby `Integer#downto`) promotes an
-    /// ancestor frame via `move_frame_to_heap` but JIT-inlined code
-    /// never does a `pop_frame` to reload r14 from the updated
-    /// `cfp.lfp` slot. The tombstone bit lets the guard catch that
-    /// case too.
-    ///
-    /// if the frame is captured, jump to *label*.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn branch_if_captured(&mut self, label: &DestLabel) {
-        monoasm! { &mut self.jit,
-            testb [r14 - (LFP_META - META_KIND)], (0b1000_1000_u8 as i8);
-            jnz label;
-        }
-    }
 }
 
 impl JitModule {
-    ///
-    /// Fetch instruction and dispatch.
-    ///
-    /// ### in
-    /// - r13: BcPc
-    ///
-    /// ### destroy
-    /// - rax, r15
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn fetch_and_dispatch(&mut self) {
-        monoasm! { &mut self.jit,
-            movq r15, (self.dispatch.as_ptr());
-            movzxb rax, [r13 + (OPECODE)]; // rax <- :0
-            addq r13, 16;
-            jmp [r15 + rax * 8];
-        };
-    }
-    /// Set outer and self for block.
-    ///
-    /// ### in
-    /// - rax: outer_lfp
-    ///
-    /// ### destroy
-    /// - rsi
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn set_block_self_outer(&mut self) {
-        self.set_block_outer();
-        monoasm! { &mut self.jit,
-            // set self
-            movq rsi, [rax - (LFP_SELF)];
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_SELF)], rsi;
-        };
-    }
-
-    /// Set outer for block.
-    ///
-    /// ### in
-    /// - rax: outer_lfp
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn set_block_outer(&mut self) {
-        monoasm! { &mut self.jit,
-            // set outer
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_OUTER)], rax;
-            // SVAR / CME are unused by block frames (they walk the
-            // outer chain to the LEP for `$~`); zero them so any stray
-            // reader sees a clean "absent" sentinel and so the GC
-            // doesn't follow uninitialised stack memory.
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_SVAR)], 0;
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_CME)], 0;
-        };
-    }
-
-    /// Set outer.
-    #[cfg(target_arch = "x86_64")]
-    fn set_method_outer(&mut self) {
-        monoasm! { &mut self.jit,
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_OUTER)], 0;
-            // Method-introducing frame: own SVAR slot starts unset
-            // (zero = "no MatchData captured"). CME also zero for
-            // now — reserved for a future migration.
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_SVAR)], 0;
-            movq [rsp - (RSP_LOCAL_FRAME + LFP_CME)], 0;
-        };
-    }
-
-    ///
-    /// Set lfp(r14) for callee.
-    ///
-    /// the local frame MUST BE on the stack.
-    #[cfg(target_arch = "x86_64")]
-    fn set_lfp(&mut self) {
-        monoasm!( &mut self.jit,
-            // set lfp
-            lea  r14, [rsp - (RSP_LOCAL_FRAME)];
-            movq [rsp - (RSP_CFP + CFP_LFP)], r14;
-        );
-    }
-
-    /// Push control frame and set outer.
-    ///
-    /// ### destroy
-    /// - rdi, rsi
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn push_frame(&mut self) {
-        monoasm!( &mut self.jit,
-            // push cfp
-            movq rdi, [rbx + (EXECUTOR_CFP)];
-            lea  rsi, [rsp - (RSP_CFP)];
-            movq [rsi], rdi;
-            movq [rbx + (EXECUTOR_CFP)], rsi;
-        );
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn restore_lfp(&mut self) {
-        monoasm!( &mut self.jit,
-            // restore lfp
-            movq r14, [rbp - (BP_CFP + CFP_LFP)];
-        );
-    }
-
-    /// Pop control frame
-    #[cfg(target_arch = "x86_64")]
-    fn pop_frame(&mut self) {
-        monoasm!( &mut self.jit,
-            // pop cfp
-            lea  r14, [rbp - (BP_CFP)];
-            movq [rbx + (EXECUTOR_CFP)], r14;
-        );
-        self.restore_lfp();
-    }
-
-    ///
-    /// Get FuncData.
-    ///
-    /// ### in
-    /// - r12: &Globals
-    /// - rdx: FuncId
-    ///
-    /// ### out
-    /// - r15: &FuncData
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn get_func_data(&mut self) {
-        monoasm! { &mut self.jit,
-            movl rdx, rdx;
-            // assumes size_of::<FuncInfo>() is 64,
-            shlq rdx, 6;
-            addq rdx, [r12 + (GLOBALS_FUNCINFO)];
-            lea  r15, [rdx + (FUNCINFO_DATA)];
-        };
-    }
-
-    ///
-    /// Get ProcData.
-    ///
-    /// ### in
-    /// - rbx: &mut Executor
-    /// - r12: &Globals
-    ///
-    /// ### out
-    /// - rax: outer_lfp: Option<LFP>
-    /// - rdx: func_id: Option<FuncId>
-    ///
-    /// ### destroy
-    /// - caller save registers
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn get_proc_data(&mut self) {
-        monoasm! { &mut self.jit,
-            movq rdi, rbx;
-            movq rsi, r12;
-            movq rax, (runtime::get_yield_data);
-            call rax;
-        }
-        self.resolve_invalidated_outer(GP::Rax);
-    }
-
-    ///
-    /// If the outer LFP in `R(reg)` points at a stack frame whose
-    /// content has already been copied to heap by `move_frame_to_heap`
-    /// (flagged `invalidated` on the tombstone), forward the pointer
-    /// to the live heap copy via the owning CFP's LFP slot.
-    ///
-    /// - `invalidated` bit = bit 3 of the `kind` byte in Meta; `kind`
-    ///   lives at `lfp - (LFP_META - META_KIND)` = `lfp - 1`.
-    /// - the LFP's CFP is at `lfp + 16`, and `cfp.lfp` slot is at
-    ///   `cfp - 8` = `lfp + 8`. Reading 8 bytes there gives the heap
-    ///   pointer that `cfp.set_lfp(heap)` stored at promotion time.
-    ///
-    /// Emits 3 instructions + 1 conditional branch on the happy
-    /// (not-invalidated) path.
-    #[cfg(target_arch = "x86_64")]
-    fn resolve_invalidated_outer(&mut self, reg: GP) {
-        let skip = self.jit.label();
-        monoasm! { &mut self.jit,
-            // Skip if outer is NULL (e.g. `get_yield_data` returned
-            // the default ProcData on a no-block-given error). The
-            // ASM callers handle that via a subsequent error check,
-            // but we must not dereference a null lfp here.
-            testq R(reg as _), R(reg as _);
-            jz   skip;
-            testb [R(reg as _) - 1], (0b0000_1000_u8 as i8);
-            jz   skip;
-            movq R(reg as _), [R(reg as _) + 8];
-        skip:
-        }
-    }
-
-    ///
-    /// ### in
-    /// - r15: &FuncData
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn call_funcdata(&mut self) -> CodePtr {
-        self.push_frame();
-        self.set_lfp();
-        monoasm! { &mut self.jit,
-            // set call site bc ptr in rcx for with_pc builtins (r13 = InlineCache = call_site + 16)
-            lea  rcx, [r13 - 16];
-            // set pc
-            movq r13, [r15 + (FUNCDATA_PC)];
-            call [r15 + (FUNCDATA_CODEPTR)];    // CALL_SITE
-        }
-        let return_addr = self.jit.get_current_address();
-        self.pop_frame();
-        return_addr
-    }
-
-    ///
-    /// Invoke the function.
-    ///
-    /// ### in
-    /// - r15: &FuncData
-    ///
-    /// ### out
-    /// - rax: return value
-    ///
-    /// ### destroy
-    /// - caller save registers
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn call_invoker(&mut self) {
-        self.push_frame();
-        self.set_lfp();
-        monoasm! { &mut self.jit,
-            // r15 : &FuncData
-            // set pc
-            movq r13, [r15 + (FUNCDATA_PC)];
-            call [r15 + (FUNCDATA_CODEPTR)];    // CALL_SITE
-            movq rdi, [rsp - (RSP_CFP)];
-            movq [rbx + (EXECUTOR_CFP)], rdi;
-        };
-    }
-
-    ///
-    /// Invoke the function.
-    ///
-    /// ### in
-    /// - r15: &FuncData
-    /// - r14: callee's Lfp
-    ///
-    /// ### out
-    /// - rax: return value
-    ///
-    /// ### destroy
-    /// - caller save registers
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn call_invoker_with_binding(&mut self) {
-        self.push_frame();
-        monoasm! { &mut self.jit,
-            // set lfp
-            movq [rsp - (RSP_CFP + CFP_LFP)], r14;
-            // r15 : &FuncData
-            // set pc
-            movq r13, [r15 + (FUNCDATA_PC)];
-            call [r15 + (FUNCDATA_CODEPTR)];    // CALL_SITE
-            movq rdi, [rsp - (RSP_CFP)];
-            movq [rbx + (EXECUTOR_CFP)], rdi;
-        };
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn push_callee_save(&mut self) {
-        monoasm! { &mut self.jit,
-            pushq r15;
-            pushq r14;
-            pushq r13;
-            pushq r12;
-            pushq rbx;
-            pushq rbp;
-        };
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn pop_callee_save(&mut self) {
-        monoasm! { &mut self.jit,
-            popq rbp;
-            popq rbx;
-            popq r12;
-            popq r13;
-            popq r14;
-            popq r15;
-        };
-    }
-
-    ///
-    /// Push stack offset for callee.
-    ///
-    /// ### in
-    /// - r15: &FuncData
-    ///
-    /// ### destoroy
-    /// - rdi
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn push_stack_offset(&mut self) {
-        monoasm! { &mut self.jit,
-            movzxw rdi, [r15 + (FUNCDATA_OFS)];
-            shlq rdi, 4;
-            addq rdi, 8;
-            subq rsp, rdi;
-            pushq rdi;
-        }
-    }
-
-    ///
-    /// Pop stack offset for callee.
-    ///
-    /// ### destoroy
-    /// - rdi
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn pop_stack_offset(&mut self) {
-        monoasm! { &mut self.jit,
-            popq rdi;
-            addq rsp, rdi;
-        }
-    }
-
-    ///
-    /// Execute GC. (for interpreter)
-    ///
-    /// ### in
-    /// - rbx: &mut Executor
-    /// - r12: &mut Globals
-    /// - r13: PC + 1
-    /// - r14: LFP
-    ///
-    /// ### out
-    /// - rax: None if Interrupt is thrown.
-    ///
-    /// ### destroy
-    /// - rax
-    /// - stack
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn vm_execute_gc(&mut self) {
-        let raise = self.entry_raise.clone();
-        self.execute_gc_inner(&raise, |_| {});
-    }
-
-    ///
-    /// Check stack overflow.
-    ///
-    /// raise StackOverFlow error if the stack pointer is below the stack limit.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn vm_check_stack(&mut self) {
-        let overflow = self.vm_stack_overflow.clone();
-        monoasm! { &mut self.jit,
-            cmpq rsp, [rbx + (EXECUTOR_STACK_LIMIT)];
-            jle overflow;
-        };
-    }
-
     ///
     /// Execute GC. (for JIT code)
     ///
@@ -1203,210 +698,6 @@ impl Codegen {
         let start1 = start1.unwrap();
         let start2 = start2.unwrap();
         (start1..start1 + size1).contains(&addr) || (start2..start2 + size2).contains(&addr)
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    icmp_main!(eq, ne, lt, le, gt, ge);
-
-    #[cfg(target_arch = "x86_64")]
-    fn icmp_teq(&mut self) {
-        self.icmp_eq()
-    }
-
-    /*///
-    /// Compare(<=>) Fixnums.
-    ///
-    /// ### in
-    /// - rdi: lhs (must be Fixnum)
-    /// - rsi: rhs (must be Fixnum)
-    ///
-    /// ### out
-    /// - rax: result(Value)
-    ///
-    /// ### destroy
-    /// - rdx
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn icmp_cmp(&mut self) {
-        monoasm! { &mut self.jit,
-            movq rax, (Value::from_ord(std::cmp::Ordering::Equal).id());
-            movq rdx, (Value::from_ord(std::cmp::Ordering::Greater).id());
-            cmpq rdi, rsi;
-            cmovgtq rax, rdx;
-            movq rdx, (Value::from_ord(std::cmp::Ordering::Less).id());
-            cmovltq rax, rdx;
-        };
-    }*/
-
-    ///
-    /// check whether lhs and rhs are fixnum.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn guard_rdi_rsi_fixnum(&mut self, generic: &DestLabel) {
-        self.guard_rdi_fixnum(generic);
-        self.guard_rsi_fixnum(generic);
-    }
-
-    ///
-    /// check whether lhs is fixnum.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn guard_rdi_fixnum(&mut self, generic: &DestLabel) {
-        monoasm!( &mut self.jit,
-            testq rdi, 0x1;
-            jz generic;
-        );
-    }
-
-    ///
-    /// check whether rhs is fixnum.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn guard_rsi_fixnum(&mut self, generic: &DestLabel) {
-        monoasm!( &mut self.jit,
-            testq rsi, 0x1;
-            jz generic;
-        );
-    }
-
-    ///
-    /// Call unary operator function.
-    ///
-    /// ### in
-    /// - rdi: receiver
-    ///
-    /// ### out
-    /// - rax: result
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn call_unop(&mut self, func: UnaryOpFn) {
-        monoasm!( &mut self.jit,
-            movq rdx, rdi;
-            movq rdi, rbx;
-            movq rsi, r12;
-            movq rax, (func as usize);
-            call rax;
-        );
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn call_binop(&mut self, func: BinaryOpFn) {
-        monoasm!( &mut self.jit,
-            movq rdx, rdi;
-            movq rcx, rsi;
-            movq rdi, rbx;
-            movq rsi, r12;
-            movq rax, (func);
-            call rax;
-        );
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn epilogue(&mut self) {
-        monoasm!( &mut self.jit,
-            leave;
-            ret;
-        );
-    }
-
-    ///
-    /// Gen code for break in block.
-    ///
-    /// rbp <- bp for a context of the outer of the block.
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn block_break(&mut self) {
-        let raise = self.entry_raise();
-        monoasm! { &mut self.jit,
-            movq rdi, rbx;
-            movq rsi, r12;
-            movq rdx, rax;
-            movq rax, (runtime::err_block_break);
-            call rax;
-            jmp  raise;
-        }
-    }
-
-    ///
-    /// Gen code for return in block.
-    ///
-    /// #### in
-    /// - rax: return value
-    /// - r13: pc + 1
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn method_return(&mut self) {
-        let raise = self.entry_raise();
-        monoasm! { &mut self.jit,
-            movq rdi, rbx;
-            movq rsi, r12;
-            movq rdx, rax;
-            movq rax, (runtime::err_method_return);
-            call rax;
-            jmp  raise;
-        }
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn method_return_specialized(&mut self, rbp_offset: usize) {
-        monoasm! { &mut self.jit,
-            lea  rbp, [rbp + (rbp_offset)];
-            leave;
-            ret;
-        }
-    }
-
-    ///
-    /// Convert Fixnum to f64.
-    ///
-    /// ### in
-    /// - R(*reg*): Value
-    ///
-    /// ### out
-    /// - xmm(*dst*)
-    ///
-    /// ### destroy
-    /// - R(*reg*)
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn integer_val_to_f64(&mut self, reg: GP, dst: u64) {
-        monoasm!(&mut self.jit,
-            sarq R(reg as _), 1;
-            cvtsi2sdq xmm(dst), R(reg as _);
-        );
-    }
-
-    ///
-    /// ### in
-    /// - r15: &FuncData
-    /// - r8: CallsiteId
-    ///
-    /// ### destroy
-    /// - caller save registers
-    ///
-    #[cfg(target_arch = "x86_64")]
-    fn generic_handle_arguments(
-        &mut self,
-        f: extern "C" fn(&mut Executor, &mut Globals, Lfp, Lfp, CallSiteId) -> Option<Value>,
-    ) {
-        monoasm! { &mut self.jit,
-            // rcx <- callee LFP
-            lea  rcx, [rsp - (RSP_LOCAL_FRAME)];
-        }
-        self.push_stack_offset();
-        monoasm! { &mut self.jit,
-            movq rdi, rbx;
-            movq rsi, r12;
-            // rdi: &mut Executor
-            // rsi: &mut Globals
-            // rdx: caller LFP
-            // rcx: callee LFP
-            // r8: CallsiteId
-            movq rax, (f);
-            movq rdx, r14;
-            call rax;
-        }
-        self.pop_stack_offset();
     }
 
     #[cfg(feature = "emit-asm")]
