+++ monoruby/src/codegen/arch/aarch64/vmgen.rs
@@ -1,742 +1,12 @@
-//! aarch64 VM backend (in progress).
+//! aarch64 VM construction (`construct_vm`), every bytecode opcode
+//! handler (`a64_op_*`), and BOP de-optimization.
 //!
-//! This is the in-crate counterpart of the x86-64 VM tier
-//! (`vmgen`/`invoker`/`wrapper` + the asm methods in `jit_module.rs` and
-//! `codegen.rs`), which is gated to `target_arch = "x86_64"`. It transcribes
-//! the x86 VM-tier shapes (dispatch core, slot access, runtime calls,
-//! conditional branches, guarded tagged-fixnum arithmetic, and the
-//! method-call frame model) into monoruby's `JitModule`/`Codegen` structure.
-//!
-//! Global-register mapping (x86-64 → aarch64), mirroring `CLAUDE.md`:
-//!
-//! | role               | x86 | aarch64 |
-//! |--------------------|-----|---------|
-//! | `&mut Executor`    | rbx | x19     |
-//! | `&mut Globals`     | r12 | x20     |
-//! | program counter    | r13 | x21     |
-//! | local frame ptr    | r14 | x22     |
-//! | accumulator        | r15 | x23     |
-//!
-//! `x19..x23` are callee-saved under AAPCS64 and are saved/restored in the VM
-//! entry prologue/epilogue (like the x86 `pushq rbp` / `leave`).
-//!
-//! See `doc/aarch64-vmgen-plan.md` for the porting plan and milestones.
-#![allow(dead_code)]
+//! Counterpart of `arch/x86_64/vmgen.rs` (+ its `vmgen/` submodules).
 
 use super::*;
 
-// ---- Global registers (mirror the x86 assignment) ----
-pub(super) const EXEC: GReg = X19; // &mut Executor   (x86 rbx)
-pub(super) const GLOBALS: GReg = X20; // &mut Globals  (x86 r12)
-pub(super) const PC: GReg = X21; // program counter   (x86 r13)
-pub(super) const LFP: GReg = X22; // local frame ptr  (x86 r14)
-pub(super) const ACC: GReg = X23; // accumulator      (x86 r15)
-// Caller-saved scratch (x9..x15).
-pub(super) const TBL: GReg = X9;
-pub(super) const OP: GReg = X10;
-pub(super) const TGT: GReg = X11;
-pub(super) const TMP: GReg = X12;
-pub(super) const TMP2: GReg = X13;
-
-impl JitModule {
-    /// Fetch the opcode and dispatch through the 256-entry table.
-    ///
-    /// x86: `movq r15,(table); movzxb rax,[r13+OPECODE]; addq r13,16;
-    /// jmp [r15+rax*8]`. On aarch64 the pc is *not* pre-advanced (A64 has no
-    /// LDUR-style negative-displacement scaled load in the builder yet), so
-    /// handlers read operands at positive offsets and advance the pc
-    /// themselves before re-dispatching.
-    ///
-    /// ### in
-    /// - x21 (PC): bytecode pointer
-    /// ### destroy
-    /// - x9 (TBL), x10 (OP), x11 (TGT)
-    pub(super) fn a64_fetch_and_dispatch(&mut self) {
-        let table = self.dispatch.as_ptr() as u64;
-        self.jit.mov_imm(TBL, table); // table base
-        self.jit.ldrb(OP, PC, OPECODE as u32); // opcode <- [pc + OPECODE]
-        self.jit.ldr_reg(TGT, TBL, OP, true); // handler <- table[opcode] (lsl #3)
-        self.jit.br(TGT);
-    }
-
-    /// Address of the local slot whose index is in `reg`:
-    /// `[r14 + reg*8 - LFP_SELF]`. x86: `negq R(r); lea R(r),[r14+R(r)*8-SELF]`.
-    /// aarch64: `neg; add Xr,x_lfp,Xr,lsl #3; sub #LFP_SELF`.
-    ///
-    /// ### in / out
-    /// - `dst`: slot index in → slot *address* out
-    pub(super) fn a64_slot_addr(&mut self, dst: GReg) {
-        self.jit.neg(dst, dst);
-        self.jit.add_lsl(dst, LFP, dst, 3);
-        self.jit.sub_imm(dst, dst, LFP_SELF as u32, 0);
-    }
-
-    /// Value of the local slot whose index is in `reg` (slot index → value).
-    pub(super) fn a64_slot_value(&mut self, dst: GReg) {
-        self.a64_slot_addr(dst);
-        self.jit.ldr(dst, dst, 0);
-    }
-
-    /// VM entry prologue: save the callee-saved global registers + fp/lr.
-    /// x86 analogue: `pushq rbp; movq rbp, rsp`.
-    pub(super) fn a64_prologue(&mut self) {
-        self.jit.stp_pre(X29, X30, SP, -16); // fp, lr
-        self.jit.stp_pre(EXEC, GLOBALS, SP, -16);
-        self.jit.stp_pre(PC, LFP, SP, -16);
-        self.jit.stp_pre(ACC, X19_PAD, SP, -16); // ACC + 8-byte pad (16-align)
-    }
-
-    /// VM exit epilogue: restore the callee-saved registers. x86: `leave; ret`.
-    pub(super) fn a64_epilogue(&mut self) {
-        self.jit.ldp_post(ACC, X19_PAD, SP, 16);
-        self.jit.ldp_post(PC, LFP, SP, 16);
-        self.jit.ldp_post(EXEC, GLOBALS, SP, 16);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-}
-
-// A scratch register used only as 16-byte-alignment padding in the prologue
-// pair stores (x24 is callee-saved but unused by the VM globals).
-const X19_PAD: GReg = X24;
-
-// ===========================================================================
-// Stub-to-link scaffolding.
-//
-// Every VM-tier method the shared code calls must exist for the aarch64
-// build to link. The methods below are placeholders that emit a `brk` trap
-// (or are no-ops) for the not-yet-ported JIT-tier surface (e.g. `f64_to_val`,
-// JIT recompile hooks); each is marked TODO(aarch64). The VM tier proper is
-// fully ported. See doc/aarch64-vmgen-plan.md.
-// ===========================================================================
-
-impl JitModule {
-    /// Bind `label` at the current position and emit a trap.
-    fn a64_brk_stub(&mut self, label: &DestLabel) {
-        self.jit.bind_label(label.clone());
-        self.jit.brk(0);
-    }
-
-    /// Like `a64_brk_stub` but reports a diagnostic id before trapping.
-    fn a64_brk_stub_diag(&mut self, label: &DestLabel, id: u64) {
-        self.jit.bind_label(label.clone());
-        self.jit.mov_imm(X0, id);
-        self.jit
-            .mov_imm(X9, crate::codegen::runtime::report_unimpl_op as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.brk(0);
-    }
-
-    /// Emit a `brk` trampoline and return its address as a typed fn pointer.
-    /// TODO(aarch64): replace each caller with the real invoker.
-    fn a64_stub_fn<T>(&mut self, id: u64) -> T {
-        let p = self.jit.get_current_address();
-        self.jit.mov_imm(X0, 0x51410000 + id); // DIAG: stub invoker hit
-        self.jit
-            .mov_imm(X9, crate::codegen::runtime::report_unimpl_op as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.brk(0);
-        // SAFETY: T is always a pointer-sized `extern "C"` fn pointer; the
-        // trampoline traps if ever called (M0 in progress).
-        unsafe { std::mem::transmute_copy::<*mut u8, T>(&p.as_ptr()) }
-    }
-
-    pub(super) fn new() -> Self {
-        let mut jit = JitMemory::new();
-        let class_version = jit.data_i32(1);
-        let bop_redefined_flags = jit.data_i32(0);
-        let const_version = jit.data_i64(1);
-        let alloc_flag = jit.data_i32(if cfg!(feature = "gc-stress") { 1 } else { 0 });
-        let pending_signals = jit.data_i32(0);
-        let entry_raise = jit.label();
-        let entry_panic = jit.label();
-        let exec_gc = jit.label();
-        let f64_to_val = jit.label();
-        let stack_overflow = jit.label();
-
-        // TODO(aarch64): emit the real entry stubs (raise/fetch_and_dispatch/
-        // panic/f64_to_val/gc). For now trap so the module links + constructs.
-        let entry_unimpl = jit.get_current_address();
-        jit.mov(X0, OP); // OP (X10) holds the opcode at dispatch time
-        jit.mov_imm(X9, crate::codegen::runtime::report_unimpl_op as *const () as u64);
-        jit.blr(X9);
-        jit.brk(0);
-        let dispatch = vec![entry_unimpl; 256];
-        let mut j = Self {
-            jit,
-            class_version,
-            const_version,
-            alloc_flag,
-            pending_signals,
-            entry_raise,
-            exec_gc,
-            f64_to_val,
-            vm_stack_overflow: stack_overflow,
-            entry_panic,
-            dispatch: dispatch.into_boxed_slice().try_into().unwrap(),
-            bop_redefined_flags,
-        };
-        let raise = j.entry_raise.clone();
-        let panic = j.entry_panic.clone();
-        let gc = j.exec_gc.clone();
-        let f64v = j.f64_to_val.clone();
-        let ovf = j.vm_stack_overflow.clone();
-        // entry_raise is emitted later in construct_vm (it needs
-        // fetch_and_dispatch); the rest stay stubs for now.
-        let _ = &raise;
-        j.a64_brk_stub_diag(&panic, 0xffff02); // entry_panic
-        // exec_gc is bound by `construct_vm` (real handler defers to
-        // entry_raise on a poll-time error).
-        let _ = gc;
-        j.a64_brk_stub_diag(&f64v, 0xffff04); // f64_to_val
-        // vm_stack_overflow is bound by `construct_vm` (the real handler
-        // needs entry_raise, which isn't emitted until then).
-        let _ = ovf;
-        j.jit.finalize();
-        j
-    }
-
-    // --- invokers / entry points ---
-    /// Real `method_invoker` for the simple case (0 args, no kw): set up the
-    /// Ruby frame and call into `vm_entry`, then return the result to Rust.
-    /// AAPCS64 in: x0 exec, x1 globals, x2 funcid, x3 self, x4 args, x5 len,
-    /// x6 block, x7 hashmap. The frame offsets match x86 because aarch64's
-    /// `stp fp,lr` (16B at vm_entry) equals x86's `call`(8B)+`push rbp`(8B).
-    /// TODO(aarch64): argument copying (assumes 0 args) + stack-overflow check.
-    /// Invoker prologue: save callee-saved globals + fp/lr + X25/X26, then set
-    /// EXEC=x0, GLOBALS=x1.
-    fn a64_invoker_prologue(&mut self) {
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.stp_pre(EXEC, GLOBALS, SP, -16);
-        self.jit.stp_pre(PC, LFP, SP, -16);
-        self.jit.stp_pre(ACC, X24, SP, -16);
-        self.jit.stp_pre(X25, X26, SP, -16);
-        self.jit.mov(EXEC, X0);
-        self.jit.mov(GLOBALS, X1);
-    }
-
-    /// fdata(X9) = funcinfo_base + (X2 = funcid)*64 + FUNCINFO_DATA.
-    fn a64_get_func_data_x2(&mut self) {
-        self.jit.lsl_imm(X10, X2, 32); // zero-extend funcid
-        self.jit.lsr_imm(X10, X10, 32);
-        self.jit.lsl_imm(X10, X10, 6); // * size_of::<FuncInfo>() (64)
-        self.jit.mov_imm(X11, GLOBALS_FUNCINFO as u64);
-        self.jit.add(X11, GLOBALS, X11);
-        self.jit.ldr(X11, X11, 0);
-        self.jit.add(X10, X10, X11);
-        self.jit.add_imm(X9, X10, FUNCINFO_DATA as u32, 0);
-    }
-
-    /// Shared invoker tail (after the frame self/block/outer/meta are set):
-    /// argument setup (simple copy or runtime arg massager), call into the
-    /// callee, then the epilogue. Assumes fdata in X9, meta in X14, x4 = args
-    /// ptr, x5 = len, x7 = kw (Option<Hashmap>).
-    fn a64_invoker_args_and_call(&mut self) {
-        let fdata = X9;
-        let generic = self.jit.label();
-        let simple_done = self.jit.label();
-        let argloop = self.jit.label();
-        let aftargs = self.jit.label();
-        let error_exit = self.jit.label();
-        // Simple iff callee is simple (Meta kind bit 4) AND len == min AND
-        // no keyword args; otherwise fall back to the runtime arg massager.
-        self.jit.lsr_imm(X10, X14, 56); // kind byte
-        self.jit.tbz_label(X10, 4, &generic);
-        self.jit.ldrh(X10, fdata, FUNCDATA_MIN as u32);
-        self.jit.cmp(X5, X10);
-        self.jit.bcond_label(Cond::Ne, &generic);
-        self.jit.cbnz_label(X7, &generic); // kw present
-        // simple copy (upward): args[0..len] -> callee slots 1..=len
-        self.jit.cbz_label(X5, &simple_done);
-        self.jit.mov(X10, X5); // down counter = len
-        self.jit.neg(X11, X5); // up counter = -len
-        self.jit.sub_imm(X12, X4, 8, 0); // args - 8
-        self.jit.bind_label(argloop.clone());
-        self.jit.add_lsl(X13, X12, X10, 3); // &args[down-1]
-        self.jit.ldr(X14, X13, 0);
-        self.jit.sub_imm(X15, SP, (RSP_LOCAL_FRAME + LFP_SELF) as u32, 0);
-        self.jit.add_lsl(X15, X15, X11, 3);
-        self.jit.str(X14, X15, 0);
-        self.jit.sub_imm(X10, X10, 1, 0);
-        self.jit.add_imm(X11, X11, 1, 0);
-        self.jit.cbnz_label(X11, &argloop);
-        self.jit.bind_label(simple_done);
-        self.jit.b_label(&aftargs);
-        // generic: handle_invoker_arguments(exec, globals, callee_lfp,
-        // arg_num, args, kw). Reserve scratch below the callee frame and
-        // preserve SP (X25) + funcdata (X26, caller-saved) across the call.
-        self.jit.bind_label(generic);
-        self.jit.mov_sp(X25, SP);
-        self.jit.mov(X26, fdata);
-        self.jit.ldrh(X10, fdata, FUNCDATA_OFS as u32);
-        self.jit.lsl_imm(X10, X10, 4);
-        self.jit.add_imm(X10, X10, 16, 0);
-        self.jit.sub(X11, X25, X10);
-        self.jit.mov_sp(SP, X11);
-        self.jit.sub_imm(X2, X25, RSP_LOCAL_FRAME as u32, 0); // callee_lfp
-        self.jit.mov(X3, X5); // arg_num = len
-        self.jit.mov(X5, X7); // kw
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        // x4 = args (unchanged); upward path => handle_invoker_arguments
-        self.jit.mov_imm(X9, runtime::handle_invoker_arguments as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.mov(X9, X26); // restore fdata
-        self.jit.mov_sp(SP, X25); // restore SP
-        self.jit.cbz_label(X0, &error_exit);
-        self.jit.bind_label(aftargs);
-        // call_invoker: push_frame
-        self.jit.ldr(X10, EXEC, EXECUTOR_CFP as u32);
-        self.jit.sub_imm(X11, SP, RSP_CFP as u32, 0);
-        self.jit.str(X10, X11, 0); // [new_cfp] = prev cfp
-        self.jit.str(X11, EXEC, EXECUTOR_CFP as u32); // exec.cfp = new cfp
-        // set_lfp
-        self.jit.sub_imm(LFP, SP, RSP_LOCAL_FRAME as u32, 0);
-        self.jit.sub_imm(X10, SP, (RSP_CFP + CFP_LFP) as u32, 0);
-        self.jit.str(LFP, X10, 0);
-        // pc = funcdata.pc; call funcdata.codeptr
-        self.jit.ldr(PC, fdata, FUNCDATA_PC as u32);
-        self.jit.ldr(X10, fdata, FUNCDATA_CODEPTR as u32);
-        self.jit.blr(X10);
-        // restore exec.cfp = [sp - RSP_CFP] (the prev cfp saved above)
-        self.jit.sub_imm(X11, SP, RSP_CFP as u32, 0);
-        self.jit.ldr(X10, X11, 0);
-        self.jit.str(X10, EXEC, EXECUTOR_CFP as u32);
-        // Converge with the result (or 0 on error) in x0; the caller emits
-        // its own epilogue (standard restore for method/block invokers, or a
-        // fiber stack-switch-back for the fiber invoker).
-        self.jit.bind_label(error_exit);
-    }
-
-    /// Standard invoker epilogue: restore the callee-saved registers saved by
-    /// a64_invoker_prologue and return (result in x0).
-    fn a64_invoker_epilogue(&mut self) {
-        self.jit.ldp_post(X25, X26, SP, 16);
-        self.jit.ldp_post(ACC, X24, SP, 16);
-        self.jit.ldp_post(PC, LFP, SP, 16);
-        self.jit.ldp_post(EXEC, GLOBALS, SP, 16);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-
-    /// Save all callee-saved GPRs (x19-x28) + fp/lr for a fiber context switch.
-    fn a64_push_callee_save(&mut self) {
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.stp_pre(X27, X28, SP, -16);
-        self.jit.stp_pre(X25, X26, SP, -16);
-        self.jit.stp_pre(X23, X24, SP, -16);
-        self.jit.stp_pre(X21, X22, SP, -16);
-        self.jit.stp_pre(X19, X20, SP, -16);
-    }
-
-    /// Restore what a64_push_callee_save saved (reverse order).
-    fn a64_pop_callee_save(&mut self) {
-        self.jit.ldp_post(X19, X20, SP, 16);
-        self.jit.ldp_post(X21, X22, SP, 16);
-        self.jit.ldp_post(X23, X24, SP, 16);
-        self.jit.ldp_post(X25, X26, SP, 16);
-        self.jit.ldp_post(X27, X28, SP, 16);
-        self.jit.ldp_post(X29, X30, SP, 16);
-    }
-
-    pub(super) fn method_invoker(&mut self) -> MethodInvoker {
-        let codeptr = self.jit.get_current_address();
-        let fdata = X9;
-        // AAPCS64 in: x0 exec, x1 globals, x2 funcid, x3 self, x4 args,
-        // x5 len, x6 block, x7 kw (Option<Hashmap>).
-        self.a64_invoker_prologue();
-        self.a64_get_func_data_x2();
-        // frame setup (method): FB = sp - (RSP_LOCAL_FRAME + LFP_SELF).
-        let fb = X12;
-        self.jit.sub_imm(fb, SP, (RSP_LOCAL_FRAME + LFP_SELF) as u32, 0);
-        self.jit.mov_imm(X13, 0);
-        self.jit.str(X3, fb, 0); // LFP_SELF  = self
-        self.jit.str(X6, fb, 8); // LFP_BLOCK = block
-        self.jit.str(X13, fb, 16); // LFP_CME   = 0
-        self.jit.str(X13, fb, 24); // LFP_SVAR  = 0
-        self.jit.ldr(X14, fdata, FUNCDATA_META as u32);
-        self.jit.str(X14, fb, 32); // LFP_META  = funcdata.meta
-        self.jit.str(X13, fb, 40); // LFP_OUTER = 0
-        self.a64_invoker_args_and_call();
-        self.a64_invoker_epilogue();
-        // SAFETY: codeptr points at an extern "C" fn with the MethodInvoker ABI.
-        unsafe { std::mem::transmute_copy::<*mut u8, MethodInvoker>(&codeptr.as_ptr()) }
-    }
-
-    pub(super) fn method_invoker2(&mut self) -> MethodInvoker2 {
-        self.a64_stub_fn(0x2)
-    }
-
-    /// Block invoker. AAPCS64 in: x0 exec, x1 globals, x2 &ProcData,
-    /// x3 self (dummy unless `with_self`), x4 args, x5 len, x6 kw.
-    /// `self` for a plain block comes from the captured outer frame.
-    /// TODO(aarch64): resolve_invalidated_outer (heap-promoted outer frames).
-    /// Set up the callee block frame from a `&ProcData` in x2 (self in x3 when
-    /// `with_self`). Produces fdata in X9 and meta in X14, and writes self /
-    /// outer / block / cme / svar / meta into the new frame. Shared by the
-    /// block and fiber invokers.
-    /// TODO(aarch64): resolve_invalidated_outer (heap-promoted outer frames).
-    fn a64_block_frame_setup(&mut self, with_self: bool) {
-        let fdata = X9;
-        if with_self {
-            // X3 = self_val (block_invoker_with_self ABI). Store at LFP_SELF.
-            // No caller-supplied block is forwarded in this variant — the
-            // LFP_BLOCK slot is zeroed below.
-            self.jit.sub_imm(X10, SP, (RSP_LOCAL_FRAME + LFP_SELF) as u32, 0);
-            self.jit.str(X3, X10, 0);
-        } else {
-            // X3 = block_val (block_invoker ABI: 4th AAPCS64 arg). Stash it
-            // into LFP_BLOCK *before* X3 is reused for outer_lfp. Without
-            // this the block passed to `Proc#call` (and through proc
-            // composition's `f.call(*a, &b)`) is silently dropped, so e.g.
-            // `(one >> two).call { |x| ... }` ran with `&arg == nil` in the
-            // inner procs.
-            self.jit.sub_imm(X10, SP, (RSP_LOCAL_FRAME + LFP_BLOCK) as u32, 0);
-            self.jit.str(X3, X10, 0);
-        }
-        // outer = [&ProcData + PROCDATA_OUTER]; func_id = [+ PROCDATA_FUNCID]
-        self.jit.ldr(X3, X2, runtime::PROCDATA_OUTER as u32); // X3 = outer lfp
-        self.jit.ldr32(X2, X2, runtime::PROCDATA_FUNCID as u32); // X2 = func_id
-        self.a64_get_func_data_x2(); // fdata = X9 (clobbers X10, X11)
-        let fb = X12;
-        self.jit.sub_imm(fb, SP, (RSP_LOCAL_FRAME + LFP_SELF) as u32, 0);
-        self.jit.mov_imm(X13, 0);
-        if !with_self {
-            // self = outer.self = [outer - LFP_SELF]
-            self.jit.sub_imm(X10, X3, LFP_SELF as u32, 0);
-            self.jit.ldr(X10, X10, 0);
-            self.jit.str(X10, fb, 0); // LFP_SELF
-            // LFP_BLOCK was already populated above with X3 (block_val).
-        } else {
-            // block_invoker_with_self has no block to forward.
-            self.jit.str(X13, fb, 8); // LFP_BLOCK = 0
-        }
-        self.jit.str(X13, fb, 16); // LFP_CME = 0
-        self.jit.str(X13, fb, 24); // LFP_SVAR = 0
-        self.jit.ldr(X14, fdata, FUNCDATA_META as u32);
-        self.jit.str(X14, fb, 32); // LFP_META
-        self.jit.str(X3, fb, 40); // LFP_OUTER = outer
-    }
-
-    fn a64_block_invoker(&mut self, with_self: bool) -> BlockInvoker {
-        let codeptr = self.jit.get_current_address();
-        self.a64_invoker_prologue();
-        self.a64_block_frame_setup(with_self);
-        self.jit.mov(X7, X6); // shared arg setup expects kw in x7
-        self.a64_invoker_args_and_call();
-        self.a64_invoker_epilogue();
-        // SAFETY: codeptr points at an extern "C" fn with the BlockInvoker ABI.
-        unsafe { std::mem::transmute_copy::<*mut u8, BlockInvoker>(&codeptr.as_ptr()) }
-    }
-
-    pub(super) fn block_invoker(&mut self) -> BlockInvoker {
-        self.a64_block_invoker(false)
-    }
-    pub(super) fn block_invoker_with_self(&mut self) -> BlockInvoker {
-        self.a64_block_invoker(true)
-    }
-
-    /// Fiber invoker (first resume). AAPCS64 in: x0 vm, x1 globals,
-    /// x2 &ProcData, x3 self, x4 args, x5 len, x6 child_vm (&mut Executor).
-    /// Switch onto the fiber's pre-allocated stack, run its block body, then
-    /// switch back to the parent on completion. (Mirrors x86 fiber_invoker.)
-    fn a64_fiber_invoker(&mut self, with_self: bool) -> FiberInvoker {
-        let codeptr = self.jit.get_current_address();
-        // switch in: save parent regs + SP, jump to the fiber's stack.
-        self.a64_push_callee_save();
-        self.jit.mov_sp(X10, SP);
-        self.jit.str(X10, X0, EXECUTOR_RSP_SAVE as u32); // parent.rsp_save = SP
-        self.jit.ldr(X10, X6, EXECUTOR_RSP_SAVE as u32);
-        self.jit.mov_sp(SP, X10); // SP = child.rsp_save (fresh stack top)
-        self.jit.str(X0, X6, EXECUTOR_PARENT_FIBER as u32); // child.parent_fiber = parent
-        self.jit.mov(EXEC, X6); // EXEC = child_vm
-        self.jit.mov(GLOBALS, X1);
-        self.a64_block_frame_setup(with_self);
-        self.jit.mov_imm(X7, 0); // fibers carry no keyword args
-        self.a64_invoker_args_and_call();
-        // body completed: mark this fiber terminated, switch back to parent.
-        self.jit.mov_imm(X10, u64::MAX); // -1 = terminated
-        self.jit.str(X10, EXEC, EXECUTOR_RSP_SAVE as u32);
-        self.jit.ldr(EXEC, EXEC, EXECUTOR_PARENT_FIBER as u32);
-        self.jit.ldr(X10, EXEC, EXECUTOR_RSP_SAVE as u32);
-        self.jit.mov_sp(SP, X10);
-        self.a64_pop_callee_save();
-        self.jit.ret();
-        // SAFETY: codeptr is an extern "C" fn with the FiberInvoker ABI.
-        unsafe { std::mem::transmute_copy::<*mut u8, FiberInvoker>(&codeptr.as_ptr()) }
-    }
-    /// Binding invoker: run a func body reusing a captured frame as its LFP
-    /// (no new local frame, no args). AAPCS64 in: x0 exec, x1 globals, x2 lfp.
-    pub(super) fn binding_invoker(&mut self) -> BindingInvoker {
-        let codeptr = self.jit.get_current_address();
-        self.a64_invoker_prologue();
-        self.jit.mov(LFP, X2); // reuse the binding's frame as LFP
-        self.jit.sub_imm(X2, LFP, LFP_FUNCID as u32, 0);
-        self.jit.ldr32(X2, X2, 0); // funcid = [lfp - LFP_FUNCID]
-        self.a64_get_func_data_x2(); // X9 = fdata
-        // push_frame; cfp.lfp = LFP (the captured frame)
-        self.jit.ldr(X10, EXEC, EXECUTOR_CFP as u32);
-        self.jit.sub_imm(X11, SP, RSP_CFP as u32, 0);
-        self.jit.str(X10, X11, 0);
-        self.jit.str(X11, EXEC, EXECUTOR_CFP as u32);
-        self.jit.sub_imm(X10, SP, (RSP_CFP + CFP_LFP) as u32, 0);
-        self.jit.str(LFP, X10, 0);
-        self.jit.ldr(PC, X9, FUNCDATA_PC as u32);
-        self.jit.ldr(X10, X9, FUNCDATA_CODEPTR as u32);
-        self.jit.blr(X10);
-        self.jit.sub_imm(X11, SP, RSP_CFP as u32, 0);
-        self.jit.ldr(X10, X11, 0);
-        self.jit.str(X10, EXEC, EXECUTOR_CFP as u32);
-        // epilogue
-        self.jit.ldp_post(X25, X26, SP, 16);
-        self.jit.ldp_post(ACC, X24, SP, 16);
-        self.jit.ldp_post(PC, LFP, SP, 16);
-        self.jit.ldp_post(EXEC, GLOBALS, SP, 16);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-        // SAFETY: codeptr is an extern "C" fn with the BindingInvoker ABI.
-        unsafe { std::mem::transmute_copy::<*mut u8, BindingInvoker>(&codeptr.as_ptr()) }
-    }
-    pub(super) fn fiber_invoker(&mut self) -> FiberInvoker {
-        self.a64_fiber_invoker(false)
-    }
-    pub(super) fn fiber_invoker_with_self(&mut self) -> FiberInvoker {
-        self.a64_fiber_invoker(true)
-    }
-
-    /// Resume a suspended fiber. In: x0 parent_vm, x1 child_vm, x2 value.
-    /// Save the parent context, switch onto the child's saved stack, and
-    /// return (the child resumes where it last yielded). x86 resume_fiber.
-    pub(super) fn resume_fiber(
-        &mut self,
-    ) -> extern "C" fn(*mut Executor, &mut Executor, Value) -> Option<Value> {
-        let codeptr = self.jit.get_current_address();
-        self.a64_push_callee_save();
-        self.jit.mov_sp(X10, SP);
-        self.jit.str(X10, X0, EXECUTOR_RSP_SAVE as u32); // parent.rsp_save = SP
-        self.jit.ldr(X10, X1, EXECUTOR_RSP_SAVE as u32);
-        self.jit.mov_sp(SP, X10); // SP = child.rsp_save
-        self.jit.str(X0, X1, EXECUTOR_PARENT_FIBER as u32); // child.parent_fiber = parent
-        self.a64_pop_callee_save();
-        self.jit.mov(X0, X2); // resume value
-        self.jit.ret();
-        // SAFETY: codeptr matches the resume_fiber ABI.
-        unsafe { std::mem::transmute_copy(&codeptr.as_ptr()) }
-    }
-
-    /// Yield from a fiber. In: x0 vm (child), x1 value. Save the child
-    /// context and switch back to the parent. x86 yield_fiber.
-    pub(super) fn yield_fiber(&mut self) -> extern "C" fn(*mut Executor, Value) -> Option<Value> {
-        let codeptr = self.jit.get_current_address();
-        self.a64_push_callee_save();
-        self.jit.mov_sp(X10, SP);
-        self.jit.str(X10, X0, EXECUTOR_RSP_SAVE as u32); // child.rsp_save = SP
-        self.jit.ldr(X0, X0, EXECUTOR_PARENT_FIBER as u32); // vm = parent
-        self.jit.ldr(X10, X0, EXECUTOR_RSP_SAVE as u32);
-        self.jit.mov_sp(SP, X10); // SP = parent.rsp_save
-        self.a64_pop_callee_save();
-        self.jit.mov(X0, X1); // yielded value
-        self.jit.ret();
-        // SAFETY: codeptr matches the yield_fiber ABI.
-        unsafe { std::mem::transmute_copy(&codeptr.as_ptr()) }
-    }
-    pub(super) fn init_stack_limit(&mut self) -> extern "C" fn(&mut Executor) -> *const u8 {
-        // executor.stack_limit = sp - MAX_STACK_SIZE (= 65536 = 16 << 12).
-        // x0 = &mut Executor (AAPCS64 arg0).
-        let codeptr = self.jit.get_current_address();
-        self.jit.mov_sp(X10, SP);
-        self.jit.sub_imm(X10, X10, 16, 1); // 16 << 12 = 65536
-        self.jit.str(X10, X0, EXECUTOR_STACK_LIMIT as u32);
-        self.jit.ret();
-        // SAFETY: codeptr is an `extern "C" fn(&mut Executor) -> *const u8`.
-        unsafe { std::mem::transmute_copy(&codeptr.as_ptr()) }
-    }
-    /// `get_class`: x0 = Value in → x0 = ClassId (u32) out. Mirrors the x86
-    /// `get_class` tag dispatch. (Invalid receivers trap rather than calling
-    /// the x86-gated `illegal_classid`.)
-    pub(super) fn get_class(&mut self) -> DestLabel {
-        let label = self.jit.label();
-        let l1 = self.jit.label();
-        let err = self.jit.label();
-        let fixnum = self.jit.label();
-        let flonum = self.jit.label();
-        let symbol = self.jit.label();
-        let nil = self.jit.label();
-        let bool_ = self.jit.label();
-        self.jit.bind_label(label.clone());
-        self.jit.tbnz_label(X0, 0, &fixnum); // bit0: fixnum
-        self.jit.tbnz_label(X0, 1, &flonum); // bit1: flonum
-        self.jit.tbnz_label(X0, 2, &l1); // bit2: other immediate
-        self.jit.cbz_label(X0, &err); // 0 is invalid
-        self.jit.ldr32(X0, X0, RVALUE_OFFSET_CLASS as u32); // heap: RValue.class
-        self.jit.ret();
-        self.jit.bind_label(l1);
-        self.jit.mov_imm(X1, 0xff);
-        self.jit.and_(X2, X0, X1);
-        self.jit.cmp_imm(X2, TAG_SYMBOL as u32, 0);
-        self.jit.bcond_label(Cond::Eq, &symbol);
-        self.jit.cmp_imm(X0, NIL_VALUE as u32, 0);
-        self.jit.bcond_label(Cond::Eq, &nil);
-        self.jit.mov_imm(X1, 8);
-        self.jit.orr(X1, X0, X1);
-        self.jit.cmp_imm(X1, TRUE_VALUE as u32, 0);
-        self.jit.bcond_label(Cond::Eq, &bool_);
-        self.jit.bind_label(err);
-        self.jit.mov_imm(X0, 0xc1a5); // DIAG: illegal_classid (get_class)
-        self.jit
-            .mov_imm(X9, crate::codegen::runtime::report_unimpl_op as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.brk(0); // TODO(aarch64): illegal_classid
-        self.jit.bind_label(fixnum);
-        self.jit.mov_imm(X0, INTEGER_CLASS.u32() as u64);
-        self.jit.ret();
-        self.jit.bind_label(flonum);
-        self.jit.mov_imm(X0, FLOAT_CLASS.u32() as u64);
-        self.jit.ret();
-        self.jit.bind_label(symbol);
-        self.jit.mov_imm(X0, SYMBOL_CLASS.u32() as u64);
-        self.jit.ret();
-        self.jit.bind_label(nil);
-        self.jit.mov_imm(X0, NIL_CLASS.u32() as u64);
-        self.jit.ret();
-        self.jit.bind_label(bool_);
-        self.jit.mov_imm(X0, BOOL_CLASS.u32() as u64);
-        self.jit.ret();
-        label
-    }
-
-    /// Per-signal async-signal-safe stub. Mirrors the x86 version
-    /// (`addl [alloc_flag], 10; orl [pending_signals], (bit); ret`). The
-    /// handler nudges `alloc_flag` (so the next GC poll notices the signal)
-    /// and OR-sets the signal bit into `pending_signals`. `signo` is folded
-    /// into the immediate bit at codegen time, so the running handler does
-    /// only memory ops and returns.
-    ///
-    /// The RMW isn't atomic (no ldxr/stxr — these would be illegal at this
-    /// level inside async-signal context too), so a nested signal during the
-    /// sequence may lose an increment / a bit. The x86 version has the same
-    /// property (no LOCK prefix); the lossy case is rare and harmless — the
-    /// next allocation poll picks it up.
-    pub(crate) fn signal_handler_for(
-        &mut self,
-        alloc_flag: DestLabel,
-        pending_signals: DestLabel,
-        signo: i32,
-    ) -> CodePtr {
-        debug_assert!((1..=32).contains(&signo));
-        let bit: u64 = 1u64 << (signo - 1);
-        let af_addr = self.jit.get_label_address(&alloc_flag).as_ptr() as u64;
-        let ps_addr = self.jit.get_label_address(&pending_signals).as_ptr() as u64;
-        let p = self.jit.get_current_address();
-        // alloc_flag += 10  (32-bit RMW)
-        self.jit.mov_imm(X0, af_addr);
-        self.jit.ldr32(X1, X0, 0);
-        self.jit.add_imm(X1, X1, 10, 0);
-        self.jit.str32(X1, X0, 0);
-        // pending_signals |= bit  (32-bit RMW)
-        self.jit.mov_imm(X0, ps_addr);
-        self.jit.ldr32(X1, X0, 0);
-        self.jit.mov_imm(X2, bit);
-        self.jit.orr(X1, X1, X2);
-        self.jit.str32(X1, X0, 0);
-        self.jit.ret();
-        p
-    }
-}
-
 impl Codegen {
-    /// entry_raise: error/exception dispatch. Calls `handle_error`, which
-    /// returns (value, dest): if `dest` is Some, resume execution there (a
-    /// rescue/ensure/retry target); otherwise unwind this VM frame and
-    /// return the error (None) to the caller (x86 `init`'s `raise:` block).
-    fn a64_gen_entry_raise(&mut self) {
-        let raise = self.entry_raise.clone();
-        let goto = self.jit.label();
-        self.jit.bind_label(raise);
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        self.jit.sub_imm(X2, LFP, LFP_META as u32, 0);
-        self.jit.ldr(X2, X2, 0); // meta = [LFP - LFP_META]
-        self.jit.mov(X3, PC); // pc = current instruction
-        self.jit
-            .mov_imm(X9, crate::codegen::jit_module::handle_error as *const () as u64);
-        self.jit.blr(X9);
-        // x0 = value (Option<Value>), x1 = dest (Option<BytecodePtr>)
-        self.jit.cbnz_label(X1, &goto);
-        // no handler: unwind this frame and return the error (x0 = None)
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-        self.jit.bind_label(goto);
-        self.jit.mov(PC, X1); // resume at the handler pc
-        self.a64_fetch_and_dispatch();
-    }
-
-    /// Bind `vm_stack_overflow`: call `stack_overflow(EXEC)` to set the
-    /// SystemStackError, then branch into `entry_raise` which unwinds the VM
-    /// frame and returns the error to the Rust caller.
-    fn a64_gen_stack_overflow(&mut self) {
-        let ovf = self.vm_stack_overflow.clone();
-        let raise = self.entry_raise.clone();
-        self.jit.bind_label(ovf);
-        self.jit.mov(X0, EXEC);
-        self.jit
-            .mov_imm(X9, super::stack_overflow as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.b_label(&raise);
-    }
-
-    /// Compare SP to `executor.stack_limit`; if SP <= limit, branch to
-    /// `vm_stack_overflow`. Uses X10 and X11 as scratch. EXEC must be valid.
-    fn a64_check_stack(&mut self) {
-        let ovf = self.vm_stack_overflow.clone();
-        self.jit.mov_sp(X10, SP);
-        self.jit.ldr(X11, EXEC, EXECUTOR_STACK_LIMIT as u32);
-        self.jit.cmp(X10, X11);
-        self.jit.bcond_label(Cond::Le, &ovf);
-    }
-
-    /// Bind `exec_gc`: call `executor::execute_gc(vm, globals)` (which also
-    /// drains pending signals). On error (returns None / X0 == 0) branch to
-    /// `entry_raise`; otherwise return to the caller. VM globals X19-X24 are
-    /// callee-saved across the Rust call so we only need to preserve fp/lr.
-    fn a64_gen_exec_gc(&mut self) {
-        let gc = self.exec_gc.clone();
-        let raise = self.entry_raise.clone();
-        self.jit.bind_label(gc);
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.mov_sp(X29, SP);
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        self.jit
-            .mov_imm(X9, crate::executor::execute_gc as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        // Option<Value>: None (0) → raise; else continue.
-        let ok = self.jit.label();
-        self.jit.cbnz_label(X0, &ok);
-        self.jit.b_label(&raise);
-        self.jit.bind_label(ok);
-        self.jit.ret();
-    }
-
-    /// VM-side GC/signal poll. If `alloc_flag >= 8` (the signal handler nudges
-    /// it by 10), call `exec_gc` which drains pending signals + runs GC. The
-    /// hot path is two loads, a compare, and a fall-through branch.
-    fn a64_vm_execute_gc(&mut self) {
-        let gc = self.exec_gc.clone();
-        let skip = self.jit.label();
-        let af_addr = self.jit.get_label_address(&self.alloc_flag).as_ptr() as u64;
-        self.jit.mov_imm(X10, af_addr);
-        self.jit.ldr32(X11, X10, 0);
-        self.jit.cmp_imm(X11, 8, 0);
-        self.jit.bcond_label(Cond::Lt, &skip);
-        self.jit.bl_label(&gc);
-        self.jit.bind_label(skip);
-    }
-
-    pub(super) fn construct_vm(&mut self) {
+    pub(in crate::codegen) fn construct_vm(&mut self) {
         self.a64_gen_entry_raise();
         self.a64_gen_stack_overflow();
         self.a64_gen_exec_gc();
@@ -975,7 +245,7 @@ impl Codegen {
 
     /// ops 121-124 `UnOp` (pos/neg/bitnot/not): fn(vm, globals, src `[pc+2]`)
     /// -> dst `[pc+4]`.
-    fn a64_op_unop(&mut self, abs: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_unop(&mut self, abs: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -990,7 +260,7 @@ impl Codegen {
 
     /// op 20 `CheckLocal`: branch by disp `[pc+0]` if local `[pc+4]` is set
     /// (non-zero); otherwise fall through (used for optional-param defaults).
-    fn a64_op_check_local(&mut self, branch: &DestLabel) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_check_local(&mut self, branch: &DestLabel) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrsw(X10, PC, 0); // disp (for the shared branch target)
         self.jit.ldrh(X12, PC, 4); // local slot
@@ -1004,7 +274,7 @@ impl Codegen {
     /// op 36 `OptCase`: dense `case`/`when` jump table. opt_case returns the
     /// branch displacement for cond slot `[pc+4]` against OptCaseId `[pc+0]`,
     /// which feeds the shared branch target.
-    fn a64_op_optcase(&mut self, branch: &DestLabel) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_optcase(&mut self, branch: &DestLabel) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldr32(X2, PC, 0); // OptCaseId
         self.jit.ldrh(X10, PC, 4); // cond slot
@@ -1023,7 +293,7 @@ impl Codegen {
     /// op 38 `Lambda`: dst `[pc+4]` <- a lambda Proc for func_id `[pc+0]`.
     /// gen_lambda may promote the current frame to the heap, so LFP is
     /// reloaded from the cfp afterward.
-    fn a64_op_lambda(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_lambda(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1047,7 +317,7 @@ impl Codegen {
     }
 
     /// op 37 `NilBr`: branch by disp `[pc+0]` if cond slot `[pc+4]` is nil.
-    fn a64_op_nilbr(&mut self, branch: &DestLabel) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_nilbr(&mut self, branch: &DestLabel) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrsw(X10, PC, 0); // disp (for shared branch target)
         self.jit.ldrh(X11, PC, 4); // cond slot
@@ -1060,7 +330,7 @@ impl Codegen {
     }
 
     /// op 23 `BlockArg`: block_arg(vm, globals, lfp, pc) -> dst `[pc+4]`.
-    fn a64_op_block_arg(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_block_arg(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1074,7 +344,7 @@ impl Codegen {
     }
 
     /// op 177 `ToA`: dst `[pc+4]` <- `to_a(src `[pc+2]`)` (splat coercion).
-    fn a64_op_to_a(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_to_a(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1090,7 +360,7 @@ impl Codegen {
     /// op 21 `BlockArgProxy`: dst `[pc+4]` <- the block handler of the frame
     /// `[pc+0]` levels up, re-encoding a proxy handler's depth. (x86
     /// `vm_block_arg_proxy`.)
-    fn a64_op_block_arg_proxy(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_block_arg_proxy(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let loop_ = self.jit.label();
         let loop_exit = self.jit.label();
@@ -1133,7 +403,7 @@ impl Codegen {
 
     /// op 148 `LoadDynVar`: dst `[pc+4]` <- the slot `[pc+2]` of the outer
     /// frame `[pc+0]` levels up the captured outer chain.
-    fn a64_op_load_dvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_load_dvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let loop_ = self.jit.label();
@@ -1168,7 +438,7 @@ impl Codegen {
 
     /// op 149 `StoreDynVar`: the slot `[pc+4]` of the outer frame `[pc+2]`
     /// levels up <- src slot `[pc+0]` of the current frame.
-    fn a64_op_store_dvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_store_dvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let loop_ = self.jit.label();
         let exit = self.jit.label();
@@ -1199,7 +469,7 @@ impl Codegen {
 
     /// op 19 `CheckKwRest`: if the kw-rest slot `[pc+4]` is nil, replace it
     /// with a fresh empty hash.
-    fn a64_op_check_kw_rest(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_check_kw_rest(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let exit = self.jit.label();
         self.jit.ldrh(X10, PC, 4);
@@ -1219,7 +489,7 @@ impl Codegen {
     }
 
     /// op 24 `CheckCvar`: check_class_var(vm, globals, name `[pc+0]`) -> dst.
-    fn a64_op_check_cvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_check_cvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1232,7 +502,7 @@ impl Codegen {
     }
 
     /// op 25 `LoadGvar`: get_global_var(vm, globals, name `[pc+0]`) -> Value.
-    fn a64_op_load_gvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_load_gvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1245,7 +515,7 @@ impl Codegen {
     }
 
     /// op 27 `LoadCvar`: get_class_var(vm, globals, name `[pc+0]`) -> Option.
-    fn a64_op_load_cvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_load_cvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1259,7 +529,7 @@ impl Codegen {
 
     /// ops 26/29 `StoreGvar`/`StoreCvar`: set_*_var(vm, globals, name `[pc+0]`,
     /// val `[pc+4]`) -> Option (error-only; no result slot).
-    fn a64_op_store_var(&mut self, set_fn: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_store_var(&mut self, set_fn: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1276,7 +546,7 @@ impl Codegen {
     }
 
     /// op 28 `AliasGvar`: alias_global_var(globals, new `[pc+0]`, old `[pc+8]`).
-    fn a64_op_alias_gvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_alias_gvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.mov(X0, GLOBALS);
         self.jit.ldr32(X1, PC, 0); // new
@@ -1290,7 +560,7 @@ impl Codegen {
 
     /// op 175 `AliasMethod`: alias_method(vm, globals, old `[pc+2]`,
     /// new `[pc+4]`).
-    fn a64_op_alias_method(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_alias_method(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1308,7 +578,7 @@ impl Codegen {
     }
 
     /// op 174 `UndefMethod`: undef_method(vm, globals, name `[pc+0]`).
-    fn a64_op_undef_method(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_undef_method(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1324,7 +594,7 @@ impl Codegen {
 
     /// op 1 `SingletonMethodDef`: `def obj.name` -- singleton_define_method(
     /// vm, globals, name `[pc+8]`, func_id `[pc+12]`, obj slot `[pc+4]`).
-    fn a64_op_singleton_method_def(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_singleton_method_def(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1343,7 +613,7 @@ impl Codegen {
 
     /// op 132 `Index`: dst[`[pc+4]`] <- base[`[pc+2]`][idx[`[pc+0]`]], with an
     /// inline ClassId cache at `[pc+8]`.
-    fn a64_op_index(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_index(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1361,7 +631,7 @@ impl Codegen {
 
     /// op 133 `IndexAssign`: base[`[pc+2]`][idx[`[pc+0]`]] <- src[`[pc+4]`],
     /// with an inline ClassId cache at `[pc+8]`.
-    fn a64_op_index_assign(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_index_assign(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1383,7 +653,7 @@ impl Codegen {
 
     /// Store the Option<Value> result in X0 to the dst slot `[pc+4]`: branch
     /// to `raise` if it is 0 (error), else store, advance PC, and dispatch.
-    fn a64_checked_store_next(&mut self, raise: &DestLabel) {
+    pub(in crate::codegen) fn a64_checked_store_next(&mut self, raise: &DestLabel) {
         let skip = self.jit.label();
         self.jit.cbz_label(X0, raise);
         self.jit.ldrh(X10, PC, 4);
@@ -1398,7 +668,7 @@ impl Codegen {
     }
 
     /// op 39 `Array`: gen_array(vm, globals, callid `[pc+0]`, &self).
-    fn a64_op_array(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_array(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1414,7 +684,7 @@ impl Codegen {
     /// op 40 `ArrayTEq`: %lhs = (%lhs === %rhs). If %lhs is an Array, returns
     /// true iff some element matches %rhs (via `===`). The result overwrites
     /// the lhs slot. Bytecode: `+0` rhs, `+2` lhs (also dst).
-    fn a64_op_array_teq(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_array_teq(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let skip = self.jit.label();
@@ -1444,7 +714,7 @@ impl Codegen {
     }
 
     /// op 176 `Hash`: gen_hash(vm, globals, src `[pc+2]`, len `[pc+0]`).
-    fn a64_op_hash(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_hash(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1460,7 +730,7 @@ impl Codegen {
 
     /// op 181 `ConcatStr` / op 86 `ConcatRegexp`: fn(vm, globals,
     /// args `[pc+2]`, len `[pc+0]`).
-    fn a64_op_concat(&mut self, abs: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_concat(&mut self, abs: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1476,7 +746,7 @@ impl Codegen {
 
     /// op 179/180 `Range`: gen_range(start `[pc+2]`, end `[pc+0]`, vm,
     /// globals, exclude_end).
-    fn a64_op_range(&mut self, exclude_end: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_range(&mut self, exclude_end: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.ldrh(X0, PC, 2);
@@ -1494,7 +764,7 @@ impl Codegen {
 
     /// op 173 `ExpandArray`: expand_array(src `[pc+4]`, &dst `[pc+2]`,
     /// len `[pc+0]`, rest `[pc+8]`). No result.
-    fn a64_op_expand_array(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_expand_array(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrh(X0, PC, 4);
         self.a64_slot_value(X0); // src (an Array Value)
@@ -1510,7 +780,7 @@ impl Codegen {
     }
 
     /// Store `X0` (a Value) into the dst slot at `[pc+4]`, advance PC, dispatch.
-    fn a64_store_dst_and_next(&mut self, skip: &DestLabel) {
+    pub(in crate::codegen) fn a64_store_dst_and_next(&mut self, skip: &DestLabel) {
         self.jit.ldrh(X10, PC, 4);
         self.jit.cbz_label(X10, skip);
         self.jit.neg(X10, X10);
@@ -1524,7 +794,7 @@ impl Codegen {
 
     /// `defined?` ops 64/69 (yield/super): fn(vm, globals) -> Value, stored to
     /// the dst slot `[pc+4]`.
-    fn a64_op_defined_to_dst(&mut self, abs: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_to_dst(&mut self, abs: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1536,7 +806,7 @@ impl Codegen {
     }
 
     /// op 65 `DefinedConst`: defined_const(vm, globals, &dst, site_id `[pc+8]`).
-    fn a64_op_defined_const(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_const(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.mov(X0, EXEC);
         self.jit.mov(X1, GLOBALS);
@@ -1552,7 +822,7 @@ impl Codegen {
 
     /// op 66 `DefinedMethod`: defined_method(vm, globals, &dst, recv `[pc+2]`,
     /// name `[pc+8]`).
-    fn a64_op_defined_method(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_method(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.mov(X0, EXEC);
         self.jit.mov(X1, GLOBALS);
@@ -1569,7 +839,7 @@ impl Codegen {
     }
 
     /// op 67 `DefinedGvar`: defined_gvar(vm, globals, name `[pc+8]`) -> Value.
-    fn a64_op_defined_gvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_gvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1582,7 +852,7 @@ impl Codegen {
     }
 
     /// op 88 `DefinedCvar`: defined_cvar(vm, globals, name `[pc+8]`) -> Value.
-    fn a64_op_defined_cvar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_cvar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.mov(X0, EXEC);
@@ -1595,7 +865,7 @@ impl Codegen {
     }
 
     /// op 68 `DefinedIvar`: defined_ivar(vm, globals, &dst, name `[pc+8]`).
-    fn a64_op_defined_ivar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_defined_ivar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.mov(X0, EXEC);
         self.jit.mov(X1, GLOBALS);
@@ -1611,7 +881,7 @@ impl Codegen {
 
     /// op 16 `LoadIvar`: slot[`[pc+4]`] <- `self.@name` (name `[pc+0]`),
     /// with an inline (ClassId, IvarId) cache at `[pc+8]`.
-    fn a64_op_load_ivar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_load_ivar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.sub_imm(X0, LFP, LFP_SELF as u32, 0);
@@ -1636,7 +906,7 @@ impl Codegen {
 
     /// op 17 `StoreIvar`: `self.@name` (name `[pc+0]`) <- slot[`[pc+4]`],
     /// with an inline (ClassId, IvarId) cache at `[pc+8]`.
-    fn a64_op_store_ivar(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_store_ivar(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -1660,7 +930,7 @@ impl Codegen {
     /// op 10 `LoadConst`: slot[`[pc+4]`] <- constant at ConstSiteId `[pc+0]`.
     /// (x86 `vm_load_const`; the JIT inline-cache slot at `[pc+8]` is not
     /// written — the VM relies on the ConstSite cache + const_version.)
-    fn a64_op_load_const(&mut self, get_fn: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_load_const(&mut self, get_fn: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let skip = self.jit.label();
@@ -1687,7 +957,7 @@ impl Codegen {
 
     /// op 11 `StoreConst`: define constant ConstSiteId `[pc+0]` <- slot
     /// `[pc+4]`, bumping const_version. (x86 `vm_store_const`.)
-    fn a64_op_store_const(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_store_const(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let cv_addr = self.jit.get_label_address(&self.const_version_label()).as_ptr() as u64;
@@ -1714,7 +984,7 @@ impl Codegen {
     /// run its body as a method with the class as `self`. Bytecode (16B):
     /// `+0` superclass slot (0 = none), `+2` base slot (0 = none),
     /// `+4` dst, `+8` name (IdentId), `+12` func_id (class body).
-    fn a64_op_class_def(&mut self, is_module: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_class_def(&mut self, is_module: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let sup_zero = self.jit.label();
@@ -1753,7 +1023,7 @@ impl Codegen {
     }
 
     /// op 22 `SingletonClassDef`: `class << base`. base = slot `[pc+0]`.
-    fn a64_op_singleton_class_def(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_singleton_class_def(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         // define_singleton_class(vm, globals, base) -> self
@@ -1773,7 +1043,7 @@ impl Codegen {
     /// to the (singleton) class, run the class body (`enter_classdef` ->
     /// call_funcdata -> `exit_classdef`) and store the result to dst `[pc+4]`.
     /// func_id is read from `[pc+12]`.
-    fn a64_class_def_run(&mut self) {
+    pub(in crate::codegen) fn a64_class_def_run(&mut self) {
         let skip = self.jit.label();
         let raise = self.entry_raise.clone();
         // enter_classdef(vm, globals, func_id, self) -> &FuncData
@@ -1853,7 +1123,7 @@ impl Codegen {
 
     /// op 2 `method_def`: `define_method(vm, globals, name, func_id)`.
     /// Bytecode: `+8` name, `+12` func_id.
-    fn a64_op_method_def(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_method_def(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.ldr32(X2, PC, 8); // name
@@ -1872,7 +1142,7 @@ impl Codegen {
     /// cache; `find_method` does the lookup every time). Handles the simple
     /// case (no kw/block/splat). Bytecode (32 bytes): `+0` callid, `+4` ret
     /// slot, `+8` pos_num, `+10` arg slot, `+12` recv slot.
-    fn a64_op_send(&mut self, is_simple: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_send(&mut self, is_simple: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         let mm = self.jit.label();
         let argloop = self.jit.label();
@@ -2048,7 +1318,7 @@ impl Codegen {
     /// ops 82/84/87 `BlockBreak`/`Retry`/`Redo`: call `f(vm[, globals, val])`
     /// to set the control-flow error, then enter entry_raise. `with_val`
     /// passes slot `[pc+4]`'s value as a 3rd argument (BlockBreak).
-    fn a64_op_err1(&mut self, abs: u64, with_val: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_err1(&mut self, abs: u64, with_val: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -2064,7 +1334,7 @@ impl Codegen {
     }
 
     /// op 83 `Raise`: raise_err(vm, exc `[pc+4]`), then enter entry_raise.
-    fn a64_op_err_raise(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_err_raise(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -2078,7 +1348,7 @@ impl Codegen {
 
     /// op 85 `EnsureEnd`: if an error is still pending after an ensure block,
     /// re-enter entry_raise; otherwise continue.
-    fn a64_op_ensure_end(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_ensure_end(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.mov(X0, EXEC);
@@ -2094,7 +1364,7 @@ impl Codegen {
     /// `+0` callid, `+4` ret slot. The block's func/outer come from
     /// `get_yield_data`; self is the block's captured self. Args are set up
     /// via the runtime arg massager (callsite-driven).
-    fn a64_op_yield(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_yield(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         let skip = self.jit.label();
@@ -2189,7 +1459,7 @@ impl Codegen {
     }
 
     /// loop_start / loop_end (ops 14/15): advance + dispatch (VM-only).
-    fn a64_op_loop(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_loop(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.add_imm(PC, PC, 16, 0);
         self.a64_fetch_and_dispatch();
@@ -2198,7 +1468,7 @@ impl Codegen {
 
     /// Unconditional branch (op 3) + the shared `branch` target used by the
     /// conditional branches. `pc += disp*16 + 16` (x86 `br_inst`/`branch:`).
-    fn a64_op_br(&mut self) -> (CodePtr, DestLabel) {
+    pub(in crate::codegen) fn a64_op_br(&mut self) -> (CodePtr, DestLabel) {
         let p = self.jit.get_current_address();
         let branch = self.jit.label();
         self.jit.ldrsw(X10, PC, 0); // disp (signed, instruction-relative)
@@ -2213,7 +1483,7 @@ impl Codegen {
     /// Conditional branch (op 4/12 `condbr`, op 5/13 `condnotbr`). Bytecode:
     /// `+0` disp (i32), `+4` cond slot. Truthiness: `(v | 0x10) != FALSE_VALUE`
     /// (both nil and false collapse to FALSE_VALUE). `not` = branch-if-falsy.
-    fn a64_op_condbr(&mut self, branch: &DestLabel, not: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_condbr(&mut self, branch: &DestLabel, not: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrsw(X10, PC, 0); // disp (kept in X10 for `branch`)
         self.jit.ldrh(X11, PC, 4); // cond slot
@@ -2231,7 +1501,7 @@ impl Codegen {
     /// Integer comparison (ops 140-146): `%dst = (%lhs <cond> %rhs)` as a Ruby
     /// boolean. Bytecode: `+0` rhs, `+2` lhs, `+4` dst. Non-fixnum traps
     /// (generic runtime fallback TODO).
-    fn a64_op_cmp(&mut self, cond: Cond, cmp_fn: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_cmp(&mut self, cond: Cond, cmp_fn: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let generic = self.jit.label();
         let skip = self.jit.label();
@@ -2270,7 +1540,7 @@ impl Codegen {
 
     /// Load the value of the slot whose (positive) index is in `idx`, into
     /// `dst`. `idx` is clobbered. (the `[r14+reg*8-LFP_SELF]` idiom.)
-    fn a64_load_slot(&mut self, idx: GReg, dst: GReg, scratch: GReg) {
+    pub(in crate::codegen) fn a64_load_slot(&mut self, idx: GReg, dst: GReg, scratch: GReg) {
         self.jit.neg(idx, idx);
         self.jit.add_lsl(scratch, LFP, idx, 3);
         self.jit.sub_imm(scratch, scratch, LFP_SELF as u32, 0);
@@ -2283,7 +1553,7 @@ impl Codegen {
     /// needed. On a Ruby error (result 0) jumps to entry_raise.
     /// ops 164-170: a binary operator with no fixnum fast path. Loads
     /// lhs `[pc+2]`, rhs `[pc+0]`, dst `[pc+4]` and calls the runtime op.
-    fn a64_op_binop(&mut self, func: BinaryOpFn) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_binop(&mut self, func: BinaryOpFn) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrh(X10, PC, 0); // rhs slot
         self.jit.ldrh(X11, PC, 2); // lhs slot
@@ -2294,7 +1564,7 @@ impl Codegen {
         p
     }
 
-    fn a64_generic_binop(&mut self, func: BinaryOpFn) {
+    pub(in crate::codegen) fn a64_generic_binop(&mut self, func: BinaryOpFn) {
         let raise = self.entry_raise.clone();
         self.jit.mov(X2, X13); // lhs
         self.jit.mov(X3, X14); // rhs
@@ -2311,7 +1581,7 @@ impl Codegen {
     /// op 162/163 `mul_rr`/`div_rr`: no fixnum fast path — straight to the
     /// runtime fallback (matches x86 `vm_binops`). Bytecode: `+0` rhs, `+2`
     /// lhs, `+4` dst.
-    fn a64_op_muldiv(&mut self, func: BinaryOpFn) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_muldiv(&mut self, func: BinaryOpFn) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrh(X10, PC, 0);
         self.jit.ldrh(X11, PC, 2);
@@ -2325,7 +1595,7 @@ impl Codegen {
     /// op 160/161 `add_rr`/`sub_rr`: fixnum fast path (`%dst = %lhs ± %rhs`)
     /// with a runtime fallback on non-fixnum/overflow. Operands are kept in
     /// X13/X14 so the fallback can use them; the result is computed in X9.
-    fn a64_op_iadd(&mut self, is_sub: bool) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_iadd(&mut self, is_sub: bool) -> CodePtr {
         let p = self.jit.get_current_address();
         let generic = self.jit.label();
         let skip = self.jit.label();
@@ -2365,7 +1635,7 @@ impl Codegen {
     /// op 172 `init_method`: allocate the method's stack frame and nil-fill the
     /// uninitialized local slots. Bytecode (relative to instruction start):
     /// `+0` stack-offset, `+2` arg_num, `+4` reg_num. (x86 `vm_init`.)
-    fn a64_op_init_method(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_init_method(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         let loop_ = self.jit.label();
@@ -2410,7 +1680,7 @@ impl Codegen {
 
     /// op 6 `immediate`: slot[`[pc+4]`] <- the immediate Value at `[pc+8]`.
     /// (x86 `vm_immediate`: `fetch_r15; movq rax,[r13-8]; vm_store_r15`.)
-    fn a64_op_immediate(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_immediate(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.ldrh(X10, PC, 4); // dst slot index
@@ -2429,7 +1699,7 @@ impl Codegen {
     /// op 7 `literal`: slot[`[pc+4]`] <- a deep copy of the literal Value at
     /// `[pc+8]`. Each evaluation yields a fresh object (mutable literals like
     /// strings/arrays). x86 `vm_literal`: `movq rdi,[r13-8]; value_deep_copy`.
-    fn a64_op_literal(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_literal(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.ldr(X0, PC, 8); // literal Value
@@ -2449,7 +1719,7 @@ impl Codegen {
 
     /// op 178 `Mov`: slot[`[pc+4]`] <- slot[`[pc+2]`]. (x86 `fetch3` +
     /// slot copy.)
-    fn a64_op_mov(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_mov(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         let skip = self.jit.label();
         self.jit.ldrh(X10, PC, 2); // src slot
@@ -2468,7 +1738,7 @@ impl Codegen {
 
     /// op 80 `ret`: return slot[`[pc+4]`]'s value (x86 `fetch_addr_r15;
     /// movq rax,[r15]; epilogue`).
-    fn a64_op_ret(&mut self) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_ret(&mut self) -> CodePtr {
         let p = self.jit.get_current_address();
         self.jit.ldrh(X10, PC, 4); // slot index
         self.jit.neg(X10, X10);
@@ -2481,186 +1751,12 @@ impl Codegen {
         self.jit.ret();
         p
     }
-
-    /// Per-method entry wrapper. For ISeq methods (VM-only build) this is the
-    /// vm-stub `b vm_entry` (x86 `gen_vm_stub`). TODO(aarch64): other
-    /// FuncKinds (Builtin/AttrReader/ConstReturn/…).
-    pub(crate) fn gen_wrapper(&mut self, globals: &Globals, fid: FuncId) -> DestLabel {
-        let entry = self.jit.label();
-        self.jit.bind_label(entry.clone());
-        match &globals.store[fid].kind {
-            FuncKind::ISeq(_) => {
-                let vm_entry = self.vm_entry();
-                self.jit.b_label(&vm_entry);
-            }
-            FuncKind::Builtin { abs_address } => {
-                let abs = *abs_address;
-                self.a64_gen_native_func_wrapper(abs);
-            }
-            FuncKind::AttrReader { ivar_name } => {
-                let name = *ivar_name;
-                self.a64_gen_attr_reader(name);
-            }
-            FuncKind::AttrWriter { ivar_name } => {
-                let name = *ivar_name;
-                self.a64_gen_attr_writer(name);
-            }
-            FuncKind::StructReader { slot_index, inline } => {
-                let (slot_index, inline) = (*slot_index, *inline);
-                self.a64_gen_struct_reader(slot_index, inline);
-            }
-            FuncKind::StructWriter { slot_index, .. } => {
-                let slot_index = *slot_index;
-                self.a64_gen_struct_writer(slot_index);
-            }
-            FuncKind::Proc(proc) => {
-                // Wrapper for a Proc converted into a method (via
-                // `define_method(:m, a_proc)`): patch the caller-set frame so
-                // that the proc's outer LFP is wired up and the meta carries
-                // PROC_METHOD_MASK (so a `return` from inside the proc behaves
-                // as a method-return rather than a non-local block exit), then
-                // tail-call the proc's codeptr.
-                let outer_ptr = match proc.outer_lfp() {
-                    Some(outer) => outer.as_ptr() as u64,
-                    None => 0,
-                };
-                let funcid = proc.func_id().get();
-                // Pre-load the outer pointer into X12 (caller-saved, but we
-                // don't call anything between this and the str below).
-                self.jit.mov_imm(X12, outer_ptr);
-                // X2 = FuncId, then resolve &FuncData into X9 (clobbers X10/X11).
-                self.jit.mov_imm(X2, funcid as u64);
-                self.a64_get_func_data_x2();
-                // [LFP - LFP_OUTER] = outer  (LFP_OUTER == 0)
-                self.jit.str(X12, LFP, 0);
-                // [LFP - LFP_META]  = funcdata.meta | PROC_METHOD_MASK
-                self.jit.ldr(X10, X9, FUNCDATA_META as u32);
-                self.jit.mov_imm(X11, Meta::PROC_METHOD_MASK);
-                self.jit.orr(X10, X10, X11);
-                self.jit.sub_imm(X13, LFP, LFP_META as u32, 0);
-                self.jit.str(X10, X13, 0);
-                // PC = funcdata.pc; tail-call funcdata.codeptr (preserve lr so
-                // the proc body returns to the original caller).
-                self.jit.ldr(PC, X9, FUNCDATA_PC as u32);
-                self.jit.ldr(X10, X9, FUNCDATA_CODEPTR as u32);
-                self.jit.br(X10);
-            }
-        }
-        self.jit.finalize();
-        entry
-    }
-
-    /// attr_reader: return `self.@ivar_name` via the cached ivar accessor.
-    fn a64_gen_attr_reader(&mut self, ivar_name: IdentId) {
-        // Per-method inline cache (InstanceVarCache, 8 bytes, init -1 = miss).
-        // Leaked on the heap so its address is known at emit time; aarch64 ADR
-        // can't reach a JIT data label from the lazily-generated wrapper.
-        let cache_addr = Box::into_raw(Box::new(-1i64)) as u64;
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.mov_sp(X29, SP);
-        self.jit.sub_imm(X0, LFP, LFP_SELF as u32, 0);
-        self.jit.ldr(X0, X0, 0); // self
-        self.jit.mov_imm(X1, ivar_name.get() as u64); // name
-        self.jit.mov(X2, GLOBALS);
-        self.jit.mov_imm(X3, cache_addr); // &cache
-        self.jit
-            .mov_imm(X9, get_instance_var_with_cache as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-
-    /// attr_writer: `self.@ivar_name = arg0` via the cached ivar setter.
-    fn a64_gen_attr_writer(&mut self, ivar_name: IdentId) {
-        let cache_addr = Box::into_raw(Box::new(-1i64)) as u64;
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.mov_sp(X29, SP);
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        self.jit.sub_imm(X2, LFP, LFP_SELF as u32, 0);
-        self.jit.ldr(X2, X2, 0); // self
-        self.jit.mov_imm(X3, ivar_name.get() as u64); // name
-        self.jit.sub_imm(X4, LFP, LFP_ARG0 as u32, 0);
-        self.jit.ldr(X4, X4, 0); // val
-        self.jit.mov_imm(X5, cache_addr); // &cache
-        self.jit
-            .mov_imm(X9, set_instance_var_with_cache as *const () as u64);
-        self.jit.blr(X9);
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-
-    /// Struct member reader: read inline slot or heap slot directly (no call).
-    fn a64_gen_struct_reader(&mut self, slot_index: u16, inline: bool) {
-        self.jit.sub_imm(X0, LFP, LFP_SELF as u32, 0);
-        self.jit.ldr(X0, X0, 0); // self
-        if inline {
-            self.jit
-                .ldr(X0, X0, (slot_index as u32) * 8 + RVALUE_OFFSET_INLINE as u32);
-        } else {
-            self.jit.ldr(X0, X0, RVALUE_OFFSET_HEAP_PTR as u32);
-            self.jit.ldr(X0, X0, (slot_index as u32) * 8);
-        }
-        self.jit.ret();
-    }
-
-    /// Struct member writer: route through set_struct_slot_with_check.
-    fn a64_gen_struct_writer(&mut self, slot_index: u16) {
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.mov_sp(X29, SP);
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        self.jit.sub_imm(X2, LFP, LFP_SELF as u32, 0);
-        self.jit.ldr(X2, X2, 0); // self
-        self.jit.sub_imm(X3, LFP, LFP_ARG0 as u32, 0);
-        self.jit.ldr(X3, X3, 0); // val
-        self.jit.mov_imm(X4, slot_index as u64);
-        self.jit.mov_imm(
-            X9,
-            crate::builtins::struct_class::set_struct_slot_with_check as *const () as u64,
-        );
-        self.jit.blr(X9);
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-
-    /// Native (builtin) method wrapper: allocate the arg stack region and call
-    /// the Rust builtin `fn(vm, globals, lfp)`. Mirrors x86
-    /// `gen_native_func_wrapper`.
-    fn a64_gen_native_func_wrapper(&mut self, abs_address: u64) {
-        self.jit.stp_pre(X29, X30, SP, -16);
-        self.jit.mov_sp(X29, SP);
-        // stack offset = ((reg_num + (RSP_LOCAL_FRAME+LFP_ARG0)/8 + 1) & !1) * 8
-        self.jit.sub_imm(X10, LFP, LFP_REGNUM as u32, 0);
-        self.jit.ldrh(X10, X10, 0); // reg_num
-        self.jit
-            .add_imm(X10, X10, ((RSP_LOCAL_FRAME + LFP_ARG0) / 8 + 1) as u32, 0);
-        self.jit.mov_imm(X11, !1u64);
-        self.jit.and_(X10, X10, X11);
-        self.jit.lsl_imm(X10, X10, 3);
-        self.jit.mov_sp(X11, SP);
-        self.jit.sub(X11, X11, X10);
-        self.jit.mov_sp(SP, X11);
-        // builtin(vm, globals, lfp)
-        self.jit.mov(X0, EXEC);
-        self.jit.mov(X1, GLOBALS);
-        self.jit.mov(X2, LFP);
-        self.jit.mov_imm(X9, abs_address);
-        self.jit.blr(X9);
-        self.jit.mov_sp(SP, X29);
-        self.jit.ldp_post(X29, X30, SP, 16);
-        self.jit.ret();
-    }
-
     /// Generic comparison handler with no fixnum fast path: calls
     /// `cmp_fn(vm, globals, lhs, rhs) -> Option<Value>` and stores the result.
     /// Used by `remove_vm_bop_optimization` to swap in the `_no_opt` runtimes
     /// after a BOP redefinition so the inline `==`/`<`/… fast paths stop being
     /// taken. Bytecode: `+0` rhs, `+2` lhs, `+4` dst (same as a64_op_cmp).
-    fn a64_op_cmp_no_opt(&mut self, cmp_fn: u64) -> CodePtr {
+    pub(in crate::codegen) fn a64_op_cmp_no_opt(&mut self, cmp_fn: u64) -> CodePtr {
         let p = self.jit.get_current_address();
         let raise = self.entry_raise.clone();
         self.jit.ldrh(X10, PC, 0); // rhs slot
@@ -2683,7 +1779,7 @@ impl Codegen {
     /// call the `_no_opt` runtimes which always invoke the redefined method
     /// instead of returning the fixnum result inline. Mirrors the x86
     /// `remove_vm_bop_optimization` in `vmgen.rs`.
-    pub(super) fn remove_vm_bop_optimization(&mut self) {
+    pub(in crate::codegen) fn remove_vm_bop_optimization(&mut self) {
         let add = self.a64_op_binop(add_values_no_opt);
         let sub = self.a64_op_binop(sub_values_no_opt);
         let mul = self.a64_op_binop(mul_values_no_opt);
