+++ monoruby/src/codegen/arch/x86_64/vmgen/init_method.rs
